Move TextInput to flat structure
