+++ packages/client/src/views/Settings/ChangePhoneModal/VerifyCodeView.tsx
@@ -18,7 +18,8 @@ import { Mutation } from 'react-apollo'
 import { changePhoneMutation } from '@client/views/Settings/mutations'
 import { get } from 'lodash'
 import { convertToMSISDN } from '@client/forms/utils'
-import { InputField, TextInput } from '@opencrvs/components/lib/forms'
+import { InputField } from '@opencrvs/components/lib/forms'
+import { TextInput } from '@opencrvs/components/lib/TextInput'
 import { useSelector, useDispatch } from 'react-redux'
 import { getUserNonce, getUserDetails } from '@client/profile/profileSelectors'
 import { EMPTY_STRING } from '@client/utils/constants'
