+++ packages/components/src/components/TextInput/index.ts
@@ -9,9 +9,4 @@
  * Copyright (C) The OpenCRVS Authors. OpenCRVS and the OpenCRVS
  * graphic logo are (registered/a) trademark(s) of Plan International.
  */
-export * from './InputField/InputField'
-export * from './InputField/InputError'
-export * from './VerifyingIndicator'
-export * from './ValidIndicator'
-export * from './PDFViewer/PDFViewer'
-export * from './InputField/InputLabel'
+export * from './TextInput'
