+++ packages/login/src/views/resetCredentialsForm/phoneNumberVerificationForm.tsx
@@ -18,7 +18,8 @@ import {
 import { authApi } from '@login/utils/authApi'
 import { phoneNumberFormat } from '@login/utils/validate'
 import { PrimaryButton } from '@opencrvs/components/lib/buttons'
-import { InputField, TextInput } from '@opencrvs/components/lib/forms'
+import { InputField } from '@opencrvs/components/lib/forms'
+import { TextInput } from '@opencrvs/components/lib/TextInput'
 import { ActionPageLight } from '@opencrvs/components/lib/interface'
 import * as React from 'react'
 import { injectIntl, WrappedComponentProps } from 'react-intl'
